# fix: Switched to JSON Copy for Performance Optimization