--- a/samtranslator/open_api/open_api.py
+++ b/samtranslator/open_api/open_api.py
@@ -581,7 +581,7 @@ class OpenApiEditor(object):
         if self.info:
             self._doc["info"] = self.info
 
-        return copy.deepcopy(self._doc)
+        return json.loads(json.dumps(self._doc))
 
     @staticmethod
     def is_valid(data):