--- a/samtranslator/open_api/open_api.py
+++ b/samtranslator/open_api/open_api.py
@@ -1,4 +1,3 @@
-import copy
 import re
 
 from samtranslator.model.intrinsics import ref